requirements.txt
azure-ai-projects==1.0.0b7
azure-ai-inference[prompts]
azure-identity
azure-search-documents==11.6.0b4
pandas
python-dotenv
opentelemetry-api
semantic-kernel==1.23.0
jupyter                                      
python-dotenv                                
notebook                                     
ipykernel                                   
pandas                                       
tabulate                                     
nbformat                                     
asyncio                                      
typing  
azure-storage-blob
aiortc==1.10.1                                     
tiktoken
azure-ai-inference[opentelemetry]
opentelemetry-sdk
opentelemetry-exporter-otlp
azure-monitor-opentelemetry